refactor: 型定義ファイルの置き場所変更
