+++ src/navigation/index.tsx
@@ -5,7 +5,7 @@ import { ColorSchemeName } from 'react-native';
 
 // from app
 import NotFoundScreen from 'app/screens/NotFoundScreen';
-import { RootStackParamList } from 'app/types';
+import { RootStackParamList } from 'app/interfaces/app';
 import BottomTabNavigator from 'app/navigation/BottomTabNavigator';
 import LinkingConfiguration from 'app/navigation/LinkingConfiguration';
